wip: refactoring out into more modular plugins
